Pass correct gc-before-benchmark parameter value to run-benchmark (#49)
benchmark* was calling run-benchmark with a 5th parameter value of
opts, which is a map in benchmark*, but was used as logical true/false
in run-benchmark, and since it was a map, is always considered as
logical true by run-benchmark.  It appears that the intent was to
instead pass the value associated with the key :gc-before-sample in
the map.